# pedidos-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de pedidos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "summary": "Retrieve a list of products",
                "description": "Retrieves a list of products based on certain criteria.",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "parameters": [
                    {
                        "name": "id_category",
                        "in": "query",
                        "description": "ID of the product category to filter by (optional)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Productos obtenidos correctamente",
                                    "data": [
                                        {
                                            "id": 1,
                                            "name": "Product 1",
                                            "price": 9.99,
                                            "description": "Description 1",
                                            "category_id": 1,
                                            "user_create_id": 1,
                                            "created_at": "2024-01-01 12:00:00",
                                            "updated_at": "2024-01-01 12:00:00"
                                        },
                                        {
                                            "id": 2,
                                            "name": "Product 2",
                                            "price": 19.99,
                                            "description": "Description 2",
                                            "category_id": 2,
                                            "user_create_id": 1,
                                            "created_at": "2024-01-02 12:00:00",
                                            "updated_at": "2024-01-02 12:00:00"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation error details"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new product",
                "description": "Creates a new product with the provided details.",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "description": "JSON object containing product details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the product",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "description": "ID of the product category",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Producto creado correctamente",
                                    "data": {
                                        "id": 1,
                                        "name": "Product Name",
                                        "price": 9.99,
                                        "description": "Product Description",
                                        "category_id": 1,
                                        "user_create_id": 1,
                                        "created_at": "2024-01-01 12:00:00",
                                        "updated_at": "2024-01-01 12:00:00"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation error details"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating the product",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Error al crear el producto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "summary": "Update an existing product",
                "description": "Updates an existing product with the provided details.",
                "operationId": "3a2800ba8920221c120204a6b425287e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON object containing product details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the product",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number"
                                    },
                                    "category_id": {
                                        "description": "ID of the product category",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Producto actualizada correctamente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation error details"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "No existe el producto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error updating the product",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Error al actualizar el producto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update an existing product",
                "description": "Updates an existing product with the provided details.",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON object containing product details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the product",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the product",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number"
                                    },
                                    "category_id": {
                                        "description": "ID of the product category",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Producto actualizada correctamente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation error details"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "No existe el producto"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error updating the product",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Error al actualizar el producto"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}